refactor: List<Number> 를 Numbers 클래스로 wrapping
