+++ src/main/java/domain/NumberGenerator.java
@@ -1,9 +1,7 @@
 package domain;
 
-import java.util.List;
-
 @FunctionalInterface
 public interface NumberGenerator {
 
-    List<Number> generate();
+    Numbers generate();
 }
